requirements.txt
discord.py
keep_alive